SPCv update + moving `build_initial_columns`
